requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
opencv-python==4.10.0.84
numpy>=1.26.0
Pillow>=10.0.0
pyzbar>=0.1.9
pytest>=8.0.0
httpx>=0.28.0
vietocr
setuptools==80.10.2
torch
torchvision
onnxruntime>=1.18.0
huggingface-hub>=0.24.0
pdfplumber>=0.11.0
ruamel.yaml>=0.18.0
cachetools>=5.3.0
pycryptodomex>=3.20.0
strenum>=0.4.15
ultralytics>=8.2.0
celery>=5.4.0
redis>=5.0.0
PyMuPDF>=1.23.0
transformers>=4.40.0
tinydb>=4.8.0
sentencepiece>=0.2.0
